Show colors set in data on legend labels
Previously, legendItems was getting the list of labels, not the total list of
objects. The colorization does "item color OR selected seed color", which was
then defaulting to the seed color.

We need to construct the legendItems using the raw data objects, so that we
get their color attribute in the same way the bars are colored.